Move all adapters to util.adapters
